+++ game-server/data/scripts/system/handlers/quest/tiamaranta_eye/_14090TracesOfThePast.java
@@ -73,7 +73,7 @@ public class _14090TracesOfThePast extends QuestHandler {
 					case SETPRO3:{
 						if (targetId == 730889) 
 							NpcActions.delete(npc);
-						QuestService.addNewSpawn(300490000, player.getInstanceId(), 802178, (float) 461.54, (float) 514.5, (float) 417, (byte) 119);
+						QuestService.addNewSpawn(300490000, player.getInstanceId(), 802178, (float) 461.54, (float) 514.5, 417, (byte) 119);
 						qs.setQuestVar(3);
 						updateQuestStatus(env);
 						return closeDialogWindow(env);
@@ -91,7 +91,7 @@ public class _14090TracesOfThePast extends QuestHandler {
 						if (targetId == 802178) 
 							NpcActions.delete(npc);
 						TeleportService2.teleportTo(player, 300500000, player.getInstanceId(), 247, 239, 124, (byte) 10);
-						QuestService.addNewSpawn(300500000, player.getInstanceId(), 205988, (float) 247.54, (float) 239.5, (float) 124, (byte) 95);
+						QuestService.addNewSpawn(300500000, player.getInstanceId(), 205988, (float) 247.54, (float) 239.5, 124, (byte) 95);
 						qs.setQuestVar(4);
 						updateQuestStatus(env);
 						return closeDialogWindow(env);
@@ -108,7 +108,7 @@ public class _14090TracesOfThePast extends QuestHandler {
 					case SETPRO5:{
 						if (targetId == 205988) 
 							NpcActions.delete(npc);
-						QuestService.addNewSpawn(300500000, player.getInstanceId(), 730890, (float) player.getX(), (float) player.getY(), (float) player.getZ(), (byte) 95);
+						QuestService.addNewSpawn(300500000, player.getInstanceId(), 730890, player.getX(), player.getY(), player.getZ(), (byte) 95);
 						qs.setQuestVar(5);
 						updateQuestStatus(env);
 						return closeDialogWindow(env);
@@ -127,7 +127,7 @@ public class _14090TracesOfThePast extends QuestHandler {
 						qs.setQuestVar(6);
 						qs.setStatus(QuestStatus.REWARD);
 						updateQuestStatus(env);
-						QuestService.addNewSpawn(300500000, player.getInstanceId(), 802178, (float) player.getX(), (float) player.getY(), (float) player.getZ(), (byte) 95);
+						QuestService.addNewSpawn(300500000, player.getInstanceId(), 802178, player.getX(), player.getY(), player.getZ(), (byte) 95);
 						return closeDialogWindow(env);
 					}
 					}
@@ -167,7 +167,7 @@ public class _14090TracesOfThePast extends QuestHandler {
 			if (player.isInsideZone(ZoneName.get("LDF4B_ITEMUSEAREA_Q14090"))) {				
 				if (var == 1) {
 					TeleportService2.teleportTo(player, 300490000, player.getInstanceId(), 504, 515, 417, (byte) 10);
-					QuestService.addNewSpawn(300490000, player.getInstanceId(), 730889, (float) 504.54, (float) 515.5, (float) 417, (byte) 95);
+					QuestService.addNewSpawn(300490000, player.getInstanceId(), 730889, (float) 504.54, (float) 515.5, 417, (byte) 95);
 					return HandlerResult.fromBoolean(useQuestItem(env, item, 1, 2, false));
 				}
 			}
